Identity dashboard refinement, testing, documentation
* setup data for dashboard, fix query/cache issue, add markdown, basic style fixes

* add seed script, enhance charts, improve error handling and messaging quality for user, and orchestrator

* finalise dashboard transparency, ui fixes on charts and setup of testing environment

* creation of frontend & backend testing, working on both, prep for next phase + rule

* add docs, update progress

+++ src-tauri/src/entities/agent_abilities.rs
@@ -13,7 +13,7 @@ pub struct Model {
     pub acquired_at: DateTimeWithTimeZone,
     pub usage_count: i32,
     pub success_count: i32,
-    pub proficiency: f32,
+    pub proficiency: f64,
     pub last_used_at: Option<DateTimeWithTimeZone>,
 }
